+++ app/controllers/dashboard_controller.rb
@@ -1,4 +1,5 @@
 class DashboardController < ApplicationController
   def show
+    @current_user = current_user
   end
 end