+++ app/controllers/journeys_controller.rb
@@ -19,17 +19,29 @@ class JourneysController < ApplicationController
     render "errors/contentful_entry_not_found", status: 500
   end
 
-  def index
-    @journeys = current_user.journeys
-  end
-
   def new
-    journey = CreateJourney.new(category_name: "catering", user: current_user).call
+    journey = CreateJourney.new(category_id: ENV["CONTENTFUL_DEFAULT_CATEGORY_ENTRY_ID"], user: current_user).call
+    
+    RecordAction.new(
+      action: "begin_journey",
+      journey_id: journey.id,
+      user_id: current_user.id,
+      contentful_category_id: ENV["CONTENTFUL_DEFAULT_CATEGORY_ENTRY_ID"]
+    ).call
     redirect_to journey_path(journey)
   end
 
   def show
     @journey = current_journey
-    @sections = @journey.sections.includes(:tasks)
+    @sections = @journey.sections.includes(:tasks).map do |section|
+      SectionPresenter.new(section)
+    end
+
+    RecordAction.new(
+      action: "view_journey",
+      journey_id: @journey.id,
+      user_id: current_user.id,
+      contentful_category_id: @journey.contentful_id
+    ).call
   end
 end