+++ app/models/task.rb
@@ -28,13 +28,8 @@ class Task < ApplicationRecord
   end
 
   def status
-    if answered_questions_count == visible_steps_count
-      return COMPLETED
-    end
-
-    if answered_questions_count > 0
-      return IN_PROGRESS
-    end
+    return COMPLETED if answered_questions_count == visible_steps_count
+    return IN_PROGRESS if answered_questions_count.positive?
 
     NOT_STARTED
   end
@@ -58,17 +53,15 @@ class Task < ApplicationRecord
     remaining_ids.first
   end
 
-  private
-
   def eager_loaded_visible_steps
     @eager_loaded_visible_steps ||= visible_steps.includes(
       %i[short_text_answer
-         long_text_answer
-         radio_answer
-         checkbox_answers
-         currency_answer
-         number_answer
-         single_date_answer]
-    )
+        long_text_answer
+        radio_answer
+        checkbox_answers
+        currency_answer
+        number_answer
+        single_date_answer]
+    ).ordered
   end
 end