+++ spec/factories/section.rb
@@ -4,5 +4,15 @@ FactoryBot.define do
     contentful_id { "5m26U35YLau4cOaJq6FXZA" }
     association :journey
     order { 0 }
+
+    trait :with_tasks do
+      transient do
+        tasks_count { 1 }
+      end
+
+      tasks do
+        Array.new(tasks_count) { association(:task) }
+      end
+    end
   end
 end